Padronizando nome do projeto
