+++ src/main/java/com/api/tcdChamados/Utils/RandomUtils.java
@@ -1,4 +1,4 @@
-package com.api.chamados.Utils;
+package com.api.tcdChamados.Utils;
 
 import java.util.Random;
