+++ src/main/java/com/api/tcdChamados/repository/ChamadoRepository.java
@@ -1,4 +1,4 @@
-package com.api.chamados.repository;
+package com.api.tcdChamados.repository;
 
 import java.util.Collection;
 
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Component;
 import org.springframework.stereotype.Repository;
 
-import com.api.chamados.model.Chamado;
+import com.api.tcdChamados.model.Chamado;
 
 @Repository
 @Component
