+++ src/main/resources/application.properties
@@ -1,3 +1,5 @@
+server.port=8083
+
 spring.application.name=chamados-service
 spring.profiles.active=prod
 spring.cloud.config.uri=http://localhost:8888
